fix: float type (#2)
`type: number`/`format: float` properties have been generated as
integer, this patch fixes them to float.

+++ app/JsonApiClientGenerator/TypeService.php
@@ -12,7 +12,7 @@ class TypeService
     {
         return match ($apiType) {
             'integer' => \Nette\PhpGenerator\Type::Int,
-            'number' => \Nette\PhpGenerator\Type::Int,
+            'number' => \Nette\PhpGenerator\Type::Float,
             'boolean' => \Nette\PhpGenerator\Type::Bool,
             'string' => match ($format) {
                 'date' => self::DATE_TIME,
